+++ EnableLidWake/EnableLidWake.hpp
@@ -59,17 +59,7 @@ private:
 	 *  @param address kinfo load address
 	 *  @param size    kinfo memory size
 	 */
-	void processKext(KernelPatcher &patcher, size_t index, mach_vm_address_t address, size_t size);
-    
-    /**
-     *  Apply kext patches for loaded kext index
-     *
-     *  @param patcher    KernelPatcher instance
-     *  @param index      kinfo index
-     *  @param patches    patch list
-     *  @param patchesNum patch number
-     */
-    void applyPatches(KernelPatcher &patcher, size_t index, const KextPatch *patches, size_t patchesNum);
+	void frameBufferPatch(KernelPatcher &patcher, size_t index, mach_vm_address_t address, size_t size);
     
     /**
      *  Current progress mask